Move pastebin_version() to version.py
This also fixes the introduced cyclic dependencies
